% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samTEMsel.main.R
\name{make_ITR}
\alias{make_ITR}
\title{make individualized treatment recommendations (ITRs) based on a \code{samTEMsel} object}
\usage{
make_ITR(samTEMsel.obj, newX = NULL, lambda.index = NULL,
  maximize = TRUE)
}
\arguments{
\item{samTEMsel.obj}{a \code{samTEMsel} object, containing the fitted models.}

\item{newX}{a (n-by-p) matrix of new values for the covariates X at which predictions are to be made; if \code{NULL}, X from the training set is used.}

\item{lambda.index}{an index of the regularization parameter \code{lambda} at which predictions are to be made; one can supply \code{lambda.opt.index} obtained from the function \code{cv.samTEMsel()}; the default is \code{NULL}, in which case the predictions are made based on the most non-sparse model.}

\item{maximize}{default is \code{TRUE}, assuming a larger value of the outcome is better; if \code{FALSE}, a smaller value is assumed to be prefered.}
}
\value{
\item{pred.new}{a (n-by-L) matrix of predicted values, with each column representing one of the L treatment options.}
\item{trt.rule}{a (n-by-1) vector of the individualized treatment recommendations}
}
\description{
The function \code{make_ITR} returns individualized treatment decision recommendations for subjects with pretreatment characteristics \code{newX}, given a \code{samTEMsel} object, \code{samTEMsel.obj}, and an (optimal) regularization parameter index, \code{lambda.index}.
}
\seealso{
\code{samTEMsel},\code{cv.samTEMsel}, \code{predict_samTEMsel}
}
\author{
Park, Petkova, Tarpey, Ogden
}
