% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samTEMsel.main.R
\name{plot_samTEMsel}
\alias{plot_samTEMsel}
\title{plot component functions from a \code{samTEMsel} object}
\usage{
plot_samTEMsel(samTEMsel.obj, newX = NULL, newA = NULL,
  scatter.plot = TRUE, lambda.index, which.index, ylims,
  solution.path = FALSE)
}
\arguments{
\item{samTEMsel.obj}{a \code{samTEMsel} object}

\item{newX}{a (n-by-p) matrix of new values for the covariates X at which plots are to be made; the default is \code{NULL}, in which case X is taken from the training set.}

\item{newA}{a (n-by-1) vector of new values for the treatment A at which plots are to be made; the default is \code{NULL}, in which case A is taken from the training set.}

\item{scatter.plot}{if \code{TRUE}, draw scatter plots of partial residuals versus the covariates; these scatter plots are made based on the training observations;  the default is \code{TRUE}.}

\item{lambda.index}{an index of the tuning parameter \code{lambda} at which plots are to be made; one can supply \code{lambda.opt.index} obtained from the function \code{cv.samTEMsel}; the default is \code{NULL}, in which case \code{plot_samTEMsel} utilizes the most non-sparse model.}

\item{which.index}{this specifies which component functions are to be plotted; the default is all p component functions, i.e., 1:p.}

\item{ylims}{this specifies the vertical range of the plots, e.g., c(-10, 10).}

\item{solution.path}{if \code{TRUE}, draw the functional norms of the fitted component functions (based on the training set) versus the regularization parameter; the default is \code{FALSE}.}
}
\description{
Produces plots of the component functions from a \code{samTEMsel} object.
}
\seealso{
\code{samTEMsel},\code{predict_samTEMsel}, \code{cv.samTEMsel}
}
\author{
Park, Petkova, Tarpey, Ogden
}
