% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samTEMsel.main.R
\name{predict_samTEMsel}
\alias{predict_samTEMsel}
\title{\code{samTEMsel} prediction function}
\usage{
predict_samTEMsel(samTEMsel.obj, newX = NULL, newA = NULL,
  type = "response", lambda.index = NULL, basis = NULL)
}
\arguments{
\item{samTEMsel.obj}{a \code{samTEMsel} object}

\item{newX}{a (n by p) matrix of new values for the covariates X at which predictions are to be made; if \code{NULL}, X from the training set is used.}

\item{newA}{a (n by 1) vector of new values for the treatment A at which predictions are to be made; if \code{NULL}, A from the training set is used.}

\item{type}{the type of prediction required; the default "response" gives the predicted responses y based on the whole model; the alternative "terms" gives the component-wise predicted responses from each of the p components (and plus the treatment-specific intercepts) of the model.}

\item{lambda.index}{an index of the tuning parameter \code{lambda} at which predictions are to be made; one can supply \code{lambda.opt.index} obtained from the function \code{cv.samTEMsel}; the default is \code{NULL}, in which case the predictions based on the most non-sparse model is returned.}

\item{basis}{a basis (design) matrix associated with the testing set (newX and newA) provided by the user; the default is \code{NULL}; this is only to efficiently implement the cross-validation in \code{cv.samTEMsel}.}
}
\value{
\item{value}{a (n-by-\code{length(lambda.index)}) matrix of predicted values; a (n-by-\code{length(lambda.index)}*(p+1)) matrix of predicted values if type =  "terms".}
}
\description{
\code{predict_samTEMsel} makes predictions given a (new) set of covariates \code{newX} and a (new) vector of treatment indicators \code{newA} based on a constrained sparse additive model \code{samTEMsel.obj}.
Specifically, \code{predict_samTEMsel} predicts the responses y based on the X-by-A interaction effect (and the A main effect) portion of the model.
}
\seealso{
\code{samTEMsel},\code{cv.samTEMsel}, \code{plot_samTEMsel}
}
\author{
Park, Petkova, Tarpey, Ogden
}
